backend/requirements.txt
# Core
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1

# LLM & Embeddings
google-generativeai==0.7.2

# Vector Database
qdrant-client==1.17.1

# HTTP client
httpx==0.27.0

# File parsing
PyPDF2==3.0.1
python-docx==1.1.0
python-multipart==0.0.9

# Rate limiting
slowapi==0.1.9

# Logging
structlog==24.1.0

# Utils
python-dotenv==1.0.1
